requirements.txt
anyio==4.9.0
APScheduler==3.6.3
blinker==1.9.0
cachetools==4.2.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
exceptiongroup==1.2.2
Flask==3.1.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pyTelegramBotAPI==4.26.0
python-telegram-bot==13.13
pytz==2025.2
requests==2.32.3
six==1.17.0
sniffio==1.3.1
tornado==6.4.2
typing_extensions==4.13.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.3.0
Werkzeug==3.1.3
zipp==3.21.0